# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WireGuard Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /nodes/{nodeId}/clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List WireGuard clients for a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Add a new WireGuard client to a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /nodes/{nodeId}/clients/{clientId}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a WireGuard client
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Remove a WireGuard client from a node
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /nodes/{nodeId}/clients/{clientId}/config:
    get:
      operationId: getClientConfig
      tags: [clients]
      summary: Get WireGuard .conf file content for a client
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Config file text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientConfig"
  /nodes/{nodeId}/clients/{clientId}/qr:
    get:
      operationId: getClientQr
      tags: [clients]
      summary: Get QR code (base64 PNG) for a client config
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: integer
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR code as base64 PNG
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientQr"
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        nodeId:
          type: integer
        name:
          type: string
        publicKey:
          type: string
        allowedIps:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nodeId
        - name
        - publicKey
        - allowedIps
        - createdAt
    ClientInput:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    ClientConfig:
      type: object
      properties:
        config:
          type: string
      required:
        - config
    ClientQr:
      type: object
      properties:
        qrBase64:
          type: string
      required:
        - qrBase64